requirements.txt
gpiozero
lgpio
python-socketio
websocket-client
hx711
eventlet